+++ app/src/main/java/com/nexters/giftarchiving/base/BaseBottomSheetFragment.kt
@@ -6,21 +6,15 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
 import androidx.databinding.ViewDataBinding
-import androidx.navigation.NavArgs
-import androidx.navigation.NavDirections
-import androidx.navigation.fragment.findNavController
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 import com.nexters.giftarchiving.BR
 import com.nexters.giftarchiving.extension.observe
 import com.nexters.giftarchiving.extension.toast
-import com.nexters.giftarchiving.util.BackDirections
-import com.nexters.giftarchiving.util.EmptyNavArgs
 
 internal abstract class BaseBottomSheetFragment<VM : BaseViewModel, VB : ViewDataBinding> :
     BottomSheetDialogFragment() {
     abstract val layoutId: Int
     abstract val viewModel: VM
-    protected open val navArgs: NavArgs = EmptyNavArgs
 
     lateinit var binding: VB
 
@@ -40,20 +34,5 @@ internal abstract class BaseBottomSheetFragment<VM : BaseViewModel, VB : ViewDat
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
         observe(viewModel.toast) { toast(it) }
-        viewModel.navArgs(navArgs)
-    }
-
-    private fun navigation(navDirections: NavDirections) {
-        if (navDirections is BackDirections) {
-            if (navDirections.destinationId == -1) {
-                findNavController().popBackStack()
-            } else {
-                with(navDirections) {
-                    findNavController().popBackStack(destinationId, inclusive)
-                }
-            }
-        } else {
-            findNavController().navigate(navDirections)
-        }
     }
 }
